do ruint arith with u32 not u64

+++ consensus/types/src/subnet_id.rs
@@ -102,11 +102,11 @@ impl SubnetId {
         spec: &ChainSpec,
     ) -> impl Iterator<Item = SubnetId> {
         // The bits of the node-id we are using to define the subnets.
-        let prefix_bits = spec.attestation_subnet_prefix_bits as u64;
+        let prefix_bits = spec.attestation_subnet_prefix_bits as u32;
 
         let node_id = U256::from_be_slice(&raw_node_id);
         // calculate the prefixes used to compute the subnet and shuffling
-        let node_id_prefix = (node_id >> (NODE_ID_BITS - prefix_bits))
+        let node_id_prefix = (node_id >> (NODE_ID_BITS as u32 - prefix_bits))
             .as_le_slice()
             .get_u64_le();
